feat: add security with jwt

+++ src/main/java/com/dios/expensesapi/service/ExpenseServiceImpl.java
@@ -6,10 +6,13 @@ import com.dios.expensesapi.exception.ResourceNotFoundException;
 import com.dios.expensesapi.mapper.ExpenseMapper;
 import com.dios.expensesapi.model.Category;
 import com.dios.expensesapi.model.Expense;
+import com.dios.expensesapi.model.User;
 import com.dios.expensesapi.repository.CategoryRepository;
 import com.dios.expensesapi.repository.ExpenseRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -21,16 +24,22 @@ public class ExpenseServiceImpl implements ExpenseService {
 
     private final ExpenseRepository expenseRepository;
     private final CategoryRepository categoryRepository;
+    private final UserService userService;
 
-    public ExpenseServiceImpl(ExpenseRepository expenseRepository, CategoryRepository categoryRepository) {
+    public ExpenseServiceImpl(ExpenseRepository expenseRepository, CategoryRepository categoryRepository, UserService userService) {
         this.expenseRepository = expenseRepository;
         this.categoryRepository = categoryRepository;
+        this.userService = userService;
     }
 
 
     @Override
     public Iterable<Expense> findAll() {
-        return expenseRepository.findAll();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userEmail = authentication.getName();
+        User currentUser = userService.findByEmail(userEmail);
+
+        return expenseRepository.findByUser(currentUser);
     }
 
     @Override
@@ -42,8 +51,13 @@ public class ExpenseServiceImpl implements ExpenseService {
     public Expense create(ExpenseDTO expenseDTO) {
         Category category = findCategory(expenseDTO);
 
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userEmail = authentication.getName();
+        User currentUser = userService.findByEmail(userEmail);
+
         try {
             Expense expense = ExpenseMapper.toEntity(expenseDTO, category);
+            expense.setUser(currentUser);
             return expenseRepository.save(expense);
         } catch (DataIntegrityViolationException e) {
             throw new DuplicateResourceException("Expense", "category", category.getName());
@@ -60,7 +74,12 @@ public class ExpenseServiceImpl implements ExpenseService {
     public Expense update(UUID id, ExpenseDTO expenseDTO) {
         Category category = categoryRepository.findById(expenseDTO.getCategoryId())
                 .orElseThrow(() -> new ResourceNotFoundException("Category"));
-        return expenseRepository.findById(id)
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userEmail = authentication.getName();
+        User currentUser = userService.findByEmail(userEmail);
+
+        return expenseRepository.findByUserAndId(currentUser, id)
                 .map(existing -> {
                     existing.setExpenseDate(expenseDTO.getExpenseDate());
                     existing.setCategory(category);
@@ -76,7 +95,17 @@ public class ExpenseServiceImpl implements ExpenseService {
         if(!expenseRepository.existsById(id)) {
             throw new ResourceNotFoundException("Expense", id.toString());
         }
-        expenseRepository.deleteById(id);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userEmail = authentication.getName();
+        User currentUser = userService.findByEmail(userEmail);
+
+        for(Expense expense : expenseRepository.findByUser(currentUser)) {
+            if(expense.getId().equals(id)) {
+                expenseRepository.deleteById(id);
+                break;
+            }
+        }
     }
 
     private Category findCategory(ExpenseDTO expenseDTO) {
